% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTargets.R
\name{all_known_species}
\alias{all_known_species}
\title{List all known species within target taxonomic group}
\usage{
all_known_species(target_taxa, target_level, tax_db, banned_words)
}
\arguments{
\item{target_taxa}{character, taxonomic group targeted by the assay - needs
to be one of "skpcofgs"}

\item{target_level}{character, taxonomic level of the targeted taxa}

\item{tax_db}{character, path to taxonomy database created with
\code{\link{make_tax_db}}}

\item{banned_words}{list, a list of words used to identify sequences
with uncertain taxonomy}
}
\value{
A table of taxonomic data
}
\description{
List all known species within target taxonomic group
}
