% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampLen.R
\name{amplicon_len}
\alias{amplicon_len}
\title{Get data on the distribution of amplicon lengths for the tested assay}
\usage{
amplicon_len(amplicon_data, target_taxa, target_level)
}
\arguments{
\item{amplicon_data}{character, amplicon data returned by bsPrimerTree.pl}

\item{target_taxa}{character, taxonomic group targeted by the assay -
needs to be one of "skpcofgs"}

\item{target_level}{character, taxonomic level of the targeted taxa}
}
\value{
A table with taxonomy and amplicon length data
}
\description{
Get data on the distribution of amplicon lengths for the tested assay
}
