% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tree.R
\name{display_tree}
\alias{display_tree}
\title{Plot the a phylogenetic tree showing amplifiable targets
from data created from find_targets}
\usage{
display_tree(output_dir, ...)
}
\arguments{
\item{output_dir}{directory where \code{\link{find_targets}} wrote output}

\item{...}{additional arguments passed to \code{\link{plot_tree_ranks}}}
}
\value{

}
\description{
Plot the a phylogenetic tree showing amplifiable targets
from data created from find_targets
}
