% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tree.R
\name{display_wordcloud}
\alias{display_wordcloud}
\title{Display a wordcloud of the taxonomic groups amplifiable}
\usage{
display_wordcloud(bsPrimerTree)
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}
}
\value{
a plot
}
\description{
Display a wordcloud of the taxonomic groups amplifiable
}
\examples{
display_wordcloud(bsPrimerTree = blastoExample)
}
