% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_data}
\alias{distance_data}
\title{Get genetic distance data on amplifiable sequences}
\usage{
distance_data(distance, target_taxa, target_level)
}
\arguments{
\item{distance}{Distance table from bsPrimerTree.pl}

\item{target_taxa}{character, taxonomic group targeted by the assay - needs
to be one of "skpcofgs"}

\item{target_level}{character, taxonomic level of the targeted taxa}
}
\value{
A summary table of taxa and distance data
}
\description{
Get genetic distance data on amplifiable sequences
}
