% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAssay.R
\name{eval_assay}
\alias{eval_assay}
\title{Evaluate metabarcoding assay}
\usage{
eval_assay(
  forward,
  reverse,
  target_taxa,
  target_level,
  assay_name,
  blast_path = "blastn",
  blast_db = "nt",
  tax_db,
  output_dir = "./output/",
  threads = 1,
  ...
)
}
\arguments{
\item{forward}{forward primer}

\item{reverse}{reverse primer}

\item{target_taxa}{taxa targeted by the assay - needs to be one of "skpcofgs"}

\item{target_level}{taxonomic level of the targeted taxa}

\item{assay_name}{name of the assay}

\item{blast_path}{path to blast version to use}

\item{blast_db}{path to blast database}

\item{tax_db}{path to taxonomy database created with make_tax_db()}

\item{output_dir}{directory to output intermediate files}

\item{threads}{number of processes to use}

\item{...}{other arguments}
}
\value{

}
\description{
Evaluate metabarcoding assay
}
\examples{

}
