% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseReblast.R
\name{get_taxa}
\alias{get_taxa}
\title{Get taxonomic data for blast results}
\usage{
get_taxa(output_dir, tax_db)
}
\arguments{
\item{output_dir}{character, directory to output intermediate files}

\item{tax_db}{character, path to taxonomy database created with
\code{\link{prepare_tax_db}}}
}
\value{
None, writes to a file
}
\description{
Get taxonomic data for blast results
}
