% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTargets.R
\name{list_all_amplifiable}
\alias{list_all_amplifiable}
\title{List all amplifiable species}
\usage{
list_all_amplifiable(output_dir, target_taxa, target_level)
}
\arguments{
\item{output_dir}{character, directory to output intermediate files}

\item{target_taxa}{character, taxonomic group targeted by the assay - needs
to be one of "skpcofgs"}

\item{target_level}{character, taxonomic level of the targeted taxa}
}
\value{
A table of taxonomic data
}
\description{
List all amplifiable species
}
