% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTargets.R
\name{list_on_target_amplifiable}
\alias{list_on_target_amplifiable}
\title{List on-target amplifiable species}
\usage{
list_on_target_amplifiable(bsPrimerTree)
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}
}
\value{
a table of taxonomic data
}
\description{
List on-target amplifiable species
}
