% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTargets.R
\name{load_find_target_data}
\alias{load_find_target_data}
\title{Load results from find_targets into output}
\usage{
load_find_target_data(bsPrimerTree, output_dir)
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}

\item{output_dir}{character, directory to output intermediate files}
}
\value{
a list
}
\description{
Load results from find_targets into output
}
