% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAssay.R
\name{missing_depends}
\alias{missing_depends}
\title{Check that the dependencies are present}
\usage{
missing_depends(blast_exe = "blastn")
}
\arguments{
\item{blast_exe}{character, path to blast executable}
}
\value{
logical, FALSE if all dependencies are present
}
\description{
Check that the dependencies are present
}
