% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampLen.R
\name{plot_amplicon_len}
\alias{plot_amplicon_len}
\title{Plot the distribution of amplicon lengths for the tested assay}
\usage{
plot_amplicon_len(bsPrimerTree)
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}
}
\description{
Plot the distribution of amplicon lengths for the tested assay
}
\examples{
\dontrun{
plot_amplicon_len(bsPrimerTree = blasto_example)
}
}
