% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{plot_distance}
\alias{plot_distance}
\title{Title}
\usage{
plot_distance(
  output_dir,
  target_taxa,
  target_level,
  levels_to_use = c("family", "genus", "species")
)
}
\arguments{
\item{levels_to_use}{}
}
\value{

}
\description{
Title
}
