% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAssay.R
\name{plot_everything}
\alias{plot_everything}
\title{Plot all the plots}
\usage{
plot_everything(bsPrimerTree)
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}
}
\value{
Several plots
}
\description{
One plot not enough? Make all the plots!
}
\examples{
\dontrun{
pdf("allThePlots.pdf", width = 10, height = 10)
plot_everything(blasto_example)
dev.off()
}
}
