% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerMismatch.R
\name{plot_primer_mismatch_locs}
\alias{plot_primer_mismatch_locs}
\title{Plot primer nucleotide mismatch locations}
\usage{
plot_primer_mismatch_locs(bsPrimerTree, target = "On-target")
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}

\item{target}{either "On-target" or "Off-target"}
}
\value{
a plot of mismatch locations
}
\description{
Plot primer nucleotide mismatch locations
}
\examples{
\dontrun{
plot_primer_mismatch_locs(bsPrimerTree = blasto_example,
                                   target = "On-target")
}
}
