% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseReblast.R
\name{potential_hits}
\alias{potential_hits}
\title{Find sequences that could have been amplified}
\usage{
potential_hits(
  output_dir,
  target_taxa,
  target_level,
  forward,
  reverse,
  banned_words
)
}
\arguments{
\item{output_dir}{character, directory to output intermediate files}

\item{target_taxa}{character, taxonomic group targeted by the assay - needs
to be one of "skpcofgs"}

\item{target_level}{character, taxonomic level of the targeted taxa}

\item{forward}{character, forward primer}

\item{reverse}{character, reverse primer}

\item{banned_words}{list, a list of words used to identify sequences
with uncertain taxonomy}
}
\description{
Find sequences that could have been amplified
}
