% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{random_alphanumeric}
\alias{random_alphanumeric}
\title{Generate a random alphanumeric string}
\usage{
random_alphanumeric(length)
}
\arguments{
\item{length}{Length of string}
}
\value{
A character string
}
\description{
Generate a random alphanumeric string
}
\examples{
random_alphanumeric(20)
}
