% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reBlast.R
\name{reblast}
\alias{reblast}
\title{Blast amplifiable sequences to assess taxonomic assignment}
\usage{
reblast(output_dir, blast_exe, blast_db, target_taxa, threads)
}
\arguments{
\item{output_dir}{character, directory to output intermediate files}

\item{blast_exe}{character, path to blast version to use}

\item{blast_db}{character, path to blast database}

\item{target_taxa}{character, taxonomic group targeted by the assay -
needs to be one of "skpcofgs"}

\item{threads}{numeric, number of processes to use}
}
\description{
Blast amplifiable sequences to assess taxonomic assignment
}
