% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAssay.R
\name{summary.bsPrimerTree}
\alias{summary.bsPrimerTree}
\title{Summarize bsPrimerTree object}
\usage{
\method{summary}{bsPrimerTree}(object)
}
\arguments{
\item{object}{A bsPrimerTree object}
}
\value{
A table of summary data from \code{\link{eval_assay}}
}
\description{
Summarize bsPrimerTree object
}
\examples{
\dontrun{
summary(blasto_example)
}
}
