% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseReblast.R
\name{tax_specificity}
\alias{tax_specificity}
\title{Quantify taxonomic specificity}
\usage{
tax_specificity(output_dir, banned_words)
}
\arguments{
\item{output_dir}{character, directory to output intermediate files}

\item{banned_words}{list, a list of words used to identify sequences
with uncertain taxonomy}
}
\description{
Quantify taxonomic specificity
}
