% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reBlast.R
\name{unique_seqs}
\alias{unique_seqs}
\title{Get unique sequences}
\usage{
unique_seqs(output_dir, target_taxa)
}
\arguments{
\item{output_dir}{character, directory to output intermediate files}

\item{target_taxa}{character, taxonomic group targeted by the assay -
needs to be one of "skpcofgs"}
}
\description{
Get unique sequences
}
