% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tree.R
\name{display_tree}
\alias{display_tree}
\title{Plot the a phylogenetic tree showing amplifiable targets
from data created from find_targets}
\usage{
display_tree(bsPrimerTree, ...)
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}

\item{...}{additional arguments passed to \code{\link{plot_tree_ranks}}}
}
\value{
a plot
}
\description{
Plot the a phylogenetic tree showing amplifiable targets
from data created from find_targets
}
\examples{
display_tree(bsPrimerTree = blastoExample)
}
