% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTargets.R
\name{find_targets}
\alias{find_targets}
\title{Title}
\usage{
find_targets(
  forward,
  reverse,
  blast_path,
  blast_db,
  tax_db,
  threads,
  output_dir
)
}
\arguments{
\item{forward}{forward primer}

\item{reverse}{reverse primer}

\item{blast_path}{path to blast version to use}

\item{blast_db}{path to blast database}

\item{tax_db}{path to taxonomy database created with make_tax_db()}

\item{threads}{number of threads to use}

\item{output_dir}{directory to write the files out to}
}
\value{

}
\description{
Title
}
\examples{

}
