% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTargets.R
\name{list_all_amplifiable}
\alias{list_all_amplifiable}
\title{List all amplifiable species}
\usage{
list_all_amplifiable(bsPrimerTree)
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}
}
\value{
A table of taxonomic data
}
\description{
List all amplifiable species
}
