% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTargets.R
\name{list_on_target_amplifiable}
\alias{list_on_target_amplifiable}
\title{Title}
\usage{
list_on_target_amplifiable(output_dir, target_taxa, target_level)
}
\arguments{
\item{target_level}{}
}
\value{

}
\description{
Title
}
