% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{plot_distance}
\alias{plot_distance}
\title{Plot genetic distance data between amplifiable sequences}
\usage{
plot_distance(
  bsPrimerTree,
  levels_to_use = c("family", "genus", "species"),
  target = "On-target"
)
}
\arguments{
\item{bsPrimerTree}{a bsPrimerTree object returned by
\code{\link{eval_assay}}}

\item{levels_to_use}{character, taxonomic levels to plot - needs
to be one of "skpcofgs"}

\item{target}{character, one of either On-target or Off-target}
}
\value{
a plot
}
\description{
Plot genetic distance data between amplifiable sequences
}
\examples{
\dontrun{
plot_distance(blasto_example)
}
}
