% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerMismatch.R
\name{plot_primer_mismatch_locs}
\alias{plot_primer_mismatch_locs}
\title{Title}
\usage{
plot_primer_mismatch_locs(
  forward,
  reverse,
  output_dir,
  target_taxa,
  target_level
)
}
\arguments{
\item{target_level}{}
}
\value{

}
\description{
Title
}
\examples{
test2 <- plot_primer_mismatch_locs(forward = "CGGTTGGGGTGACCTCGGA",
                                   reverse = "GCTGTTATCCCTAGGGTAACT",
                                   output_dir = "C:/Users/MattDellLaptop/Desktop/mam16SOut2/",
                                   target_taxa = "Mammalia",
                                   target_level = "class")
}
