% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerMismatch.R
\name{primer_mismatch_count}
\alias{primer_mismatch_count}
\title{Add info to primer nucleotide mismatches table}
\usage{
primer_mismatch_count(mismatch_data, target_taxa, target_level)
}
\arguments{
\item{mismatch_data}{mismatch data output by bsPrimerTree.pl}

\item{target_level}{}
}
\value{
A table of primer mismatch data
}
\description{
Add info to primer nucleotide mismatches table
}
