% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerMismatch.R
\name{primer_mismatch_count}
\alias{primer_mismatch_count}
\title{Title}
\usage{
primer_mismatch_count(output_dir, target_taxa, target_level)
}
\arguments{
\item{target_level}{}
}
\value{

}
\description{
Title
}
