% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerMismatch.R
\name{primer_mismatch_loc}
\alias{primer_mismatch_loc}
\title{Title}
\usage{
primer_mismatch_loc(output_dir, target_taxa, target_level)
}
\arguments{
\item{target_level}{}
}
\value{

}
\description{
Title
}
\examples{
test <- primer_mismatch_loc(output_dir = "C:/Users/MattDellLaptop/Desktop/mam16SOut2/",
                            target_taxa = "Mammalia",
                            target_level = "class")
}
