% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerMismatch.R
\name{primer_mismatch_loc}
\alias{primer_mismatch_loc}
\title{Find location of primer nucleotide mismatches}
\usage{
primer_mismatch_loc(output_dir, target_taxa, target_level)
}
\arguments{
\item{output_dir}{character, directory to output intermediate files}

\item{target_taxa}{character, taxonomic group targeted by the assay -
needs to be one of "skpcofgs"}

\item{target_level}{character, taxonomic level of the targeted taxa}
}
\value{
A table of primer mismatch location data
}
\description{
Find location of primer nucleotide mismatches
}
